[FIX] sign: translate ecological report + fix English
The cards in the ecological report weren't translated due to the strings
being defined in t-values. This is switched so that they will now be
correctly exported to the .pot file + translated.

Additionally cleaned up some of the incorrect English for a better UX.
More terms could be improved/fixed as well, but they do not affect the
users as much + are already translated so changing them was avoided in
stable.

- Carbon => "carbon emissions": this or CO_2 are used when referring
  to reducing pollution due to burning fuel. "carbon" on its own implies
  the element, whereas adding "emissions" to the end of it implies CO_2
  (emissions), because English ¯\_(ツ)_/¯
- sparred => saved/reduced: sparred is both spelled wrong (i.e. should
  be spared) and incorrect to use in this case
- like => that's: "like" doesn't make sense to use in this case and is
  confusing to read. "That's" indicates how much the user is
  saving/reducing/etc. "Approximately" or "about" also would have
  worked, but are less friendly sounding

closes odoo/enterprise#55894

X-original-commit: d7b95c57026b64528dda409d6112d279e2becc8e
Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>