[IMP] planning: Remove default working time and hide data in list view
Currently, a working calendar is set by default on material resources
which doesn't make a lot of sense as tools and machines don't have the
same capacity limitations and hours restrictions as human resources do.

Remove the default value for the working time. It need to be done at 2
places for the folowing reasons:
    - `default_get`: When creating a shift on the planning, when we
        create a new resource, it need to be set without the default
        working time.
    - `_onchange_company_id`: When creating a resource from the
        planning material list view, the `_onchange_company_id`
        function from the resource module  is called and don't
        take into account that the `default_get` already changed the
        value, so we need to override it to ensure it won't reset the
        `calendar_id`.

task-3692156

closes odoo/enterprise#54692

Related: odoo/odoo#150127
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>