[IMP] knowledge: store comment anchors
This commit adds the storage of text anchors for comments.
Before this commit we didn't store the text's anchors and displayed it
dynamically.
The advantage of this approach is that we didn't need to store any text inside
the DB for each comment => this could've led to big sums of data for not a big
gain.
The disadvantage is that if the anchor doesn't exist anymore and we lose the
context of the comment.

Now, a new field `article_anchor_text` stores up to 1200 characters of the
anchor's text. When the anchor is longer than 50 chars we display a Read More
button that will display the anchor's stored text.

task-3672422

Part-of: odoo/enterprise#54716