[FIX] appointment_hr: get employee even for unsynced partners
Partner employees are usually synced with that of the user
if the partner represents a user. However it's possible
to reach a state where the partner and the user have different
employees. In that case we want to use the employee of the
user, as is the case when computing availability for appointments
on the website.

task-3452277

X-original-commit: c070e7424a72cab90bca0b26fa7245e5f7cc0a02
Part-of: odoo/enterprise#54916