[I18N] l10n_mx_hr_payroll: no unicode quotes in payroll report name
Thanks to the safe_eval call

https://github.com/odoo/enterprise/blob/0ae2e56fc3e14ee0afdf314fda5c028ba9d33242/hr_payroll/controllers/main.py#L43

in `get_payroll_report_print`, whenever a non-ASCII character is passed,
a Python `SyntaxError: invalid character` will occur because this
character will be passed to `compile`

https://github.com/odoo/odoo/blob/e7a511b7b5af57696e691f6f0a71bec63053f860/odoo/tools/safe_eval.py#L216

which expects only valid python code. Therefore we remove the
unneccessary non-ASCII quote from the .po file that was probably a
mistake in the first place.

opw-3688140

closes odoo/enterprise#54729

Signed-off-by: Louis Wicket (wil) <wil@odoo.com>