[FIX] l10n_lt_reports: Balance Sheet: respect legal line order
"Total Assets" and "Total Equity and Liabilities" were inverted. This occured because those lines used to declare an explicit sequence in 15.0, and this sequence got removed from the data file in 16.0. We fix that by enforcing a sequence, and also reorder the lines in the file in order to ease the forward-port to 17.0, where the auto-sequencing feature is used (this way, we also minimize the number of conflicts when forward-porting a fix in this report in the future).

OPW 3608139

X-original-commit: e2d56f58280b0edba0583dc777384b409c3242a5
Part-of: odoo/enterprise#54777