[FIX] sale_amazon: don't update FBA stock inventory
Following commit 1b2977811516cce4830ceaf1a7585c8f7061243a we now send the stock inventory for all
the products linked to an Amazon offer. At the time, it was decided to send FBA and FBM updates at
the same time because:
- The Amazon documentation didn't say anything about it
- It avoided the problem of a strange configuration of the customer where a product in FBM could be
placed in the Amazon location, as nothing ensures that everything placed there is, indeed, a FBA
product, even if this is the recommended setting.
- After testing the synchronization of the FBM stock, we couldn't see any changes in the Amazon
backend, which lead us to think that we couldn't, in fact, update the stock of a FBA product as
Amazon would ensure the stock used would be the one in their warehouse.

Recently, we got some problem with FBA order that Amazon considered being FBM. After investigation,
it was found that the stock send for FBA would be set as having a "DEFAULT" fulfilment channel,
but this wasn't shown anywhere on Amazon backend, only in their DB. So from time to time, instead
of using the set fulfilment channel of a product, Amazon would use that default one, with a
separate, virtual, stock.

With this fix, we ensure that this "default" channel is now filled with a quantity of 0 instead,
for FBA product. We assume the set-up of the Amazon account on Odoo is correctly made and all
products in the Amazon location are FBA.

task-3693853

closes odoo/enterprise#55909

X-original-commit: b37cacd0b182e2fd6f90eee668f81de1d26e5b11
Signed-off-by: Morgane Demesmaeker <edm@odoo.com>