[FIX] web_studio: reportEditor: spawn ImageCrop in an overlay
Before this commit, spawning the cropping tool from the Wysiwyg's toolbar
made it appear within the iframe, without any coherence whatsoever regarding the display
or usability.

After this commit, we manually spawn the image crop tool in an overlay of the main window.
It is much clearer UX wise and cleaner technically.

opw-3484065

closes odoo/enterprise#55049

X-original-commit: 0b2a55c24773d2dbc0868255649ac00c43f4702e
Related: odoo/odoo#150904
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>