[FIX] web_studio: reportEditor: spawn ImageCrop in an overlay
Before this commit, spawning the cropping tool from the Wysiwyg's toolbar
made it appear within the iframe, without any coherence whatsoever regarding the display
or usability.

After this commit, we manually spawn the image crop tool in an overlay of the main window.
It is much clearer UX wise and cleaner technically.

opw-3484065

closes odoo/enterprise#55095

X-original-commit: b50ffc3326e0805a8eff39954b4499b458f772e1
Related: odoo/odoo#151002
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>