[FIX] stock_barcode: increment right qty
When a product tracked by lots is reserved and multiple lots are
reserved, in the Barcode App, each line has a "+ qty" button where the
quantity is the remaining quantity.
That said, this remaining quantity was always computed regarding the
difference between the move line's done and reserved quantity which
means if the user scans a not reserved lot, those buttons aren't
reliable anymore.

How to reproduce:
- Create a product tracked by lots and set a barcode;
- Update the quantity on hand for this product like following:
  - 3x lot1
  - 3x lot2
  - 3x lot3
- Create a delivery for 6x this product and confirm it;
- Open the delivery in the Barcode App, unfold the line for the tracked
  product, you should have:
  - 0/3 lot1, [+3]
  - 0/3 lot2, [+3]
- Scan lot1 one time and lot3 three times, you now have:
  - 3 lot3, [+1]
  - 1/3 lot1, [+2]
  - 0/3 lot2, [+2]
-> Since the total demand is 6 and 4 quantities have been processed, no
button should display a quantity greater than 2 since it's the
remaining quantity.

To fix that, the quantity used by the increment button is also computed
regarding the parent line's quantity.

task-3688770

X-original-commit: b7864cf657603969f58cf7521def220117fbb9db
Part-of: odoo/enterprise#67649
Related: odoo/odoo#175035
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>
Signed-off-by: Steve Van Essche <svs@odoo.com>