[FIX] timesheet_grid: avoid extra condition for group expand in grid view
Before this commit, when the group expand is triggered in the grid view of
timesheets to get the empty lines according to the timesheets recorded in the
previous period, the domain to gather the additional lines will make sure the
projects still have the timesheets feature enabled and the tasks are always
actives.

This commit removes those extra conditions because we could assume in one week
the project in which we previously record a timesheet, still has the timesheets
feature enabled. Same idea for the tasks, we could also assume the tasks are
still actives. For the reason about removing those extra conditions is because
`('task_id.active', '=', True)` leaf will also trigger the `ir.rule` to make
sure the user can read the fields in the task, idem for project with
`('project_id.allow_timesheets', '=', True)` leaf.

X-original-commit: e4e3b8cc0cd690a1ee68161da710ed116f94c89d
Part-of: odoo/enterprise#55261