[FIX] knowledge: fix access error when restoring article
Currently, the user encounters an access error when attempting to restore
an article whose parent is inaccessible. Upon restoration, the system
accesses the parent article to determine whether the article to restore
should be detached from its parent.

Note: If the parent article is in the trash, the article should be
detached from its parent. Otherwise, it will be automatically deleted
when deleting the parent article due to the cascading delete.

When the user does not have access to the parent article, the system will
raise an access error when doing the check. This commit addresses the
issue by introducing a `sudo` on the read call causing the access error.

Steps to reproduce the issue:
1. Log in as "Mitchell Admin"
2. Create an article in the workspace
3. Add "Marc Demo" to the member list with "No access"
4. Create an article under the article create at step 2
5. Set the permission of "Marc Demo" to "Can write"
6. In the "More option" panel, click on "Delete"
7. Log in as "Marc Demo"
8. On the Knowledge editor, click on "Open the Trash"
9. Open the article you deleted
10. In the "More option" panel, click on "Restore From Trash"

=> "Marc Demo" gets an access error when restoring the article.

TO BE: "Marc Demo" should be able to restore the article and should not
get any access error.

task-3648854

closes odoo/enterprise#54892

X-original-commit: 60742ff55ec444b0834b761a7b508791994f7e00
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>