[FIX] *: fix related field search used in domain
In the community branch, we changed the behavior of the search method of related field.
`'resource_type', '!=', 'material'` was filterout record where `resource_id IS NULL`. It is not the case anymore, then to pass the `test_performance_recompute_shifts_in_leave_periods` test, we need to `('resource_id', '!=', False)`

Part-of: odoo/enterprise#46795