[FIX] l10n_pt_reports: Balance Sheet rework
In Portugal, there are four regimes of regulations, which include a CoA
and financial statements: general regime, small companies,
micro-companies and non-profits. Regulations are published at
https://www.cnc.min-financas.pt/snc2016.html

As a result of the refactor in https://github.com/odoo/odoo/pull/87572,
the CoA (almost) perfectly follows the general regime regulation.
However, the financial reports are not up-to-date.

This commit implements the Portuguese Balance Sheet and Profit and Loss
for companies under the general regime, as defined at pp.42-45, 53-55 of
https://www.occ.pt/fotos/editor2/manualapoiosaf-t_1.pdf.

The correspondence between account 'Taxonomy Codes' and account codes
was taken from
https://www.occ.pt/fotos/editor2/taxonomiasplanocontas_fev2019.pdf.

taskid:3060790

closes odoo/enterprise#58709

X-original-commit: f059edcf396fe7bc0c4fb71aeee468cefbc04e4d
Related: odoo/odoo#157780
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>