[FIX] documents : adjust dropdown interfered display in documents module
Steps to reproduce:
	1- Install Documents module
	2- Go to documents list view
	3- Click on the dropdown menu for additional fields selection

Current behavior before PR:
The dropdown menu for additional fields selection in documents list
view is interfered with the table cells so the user will not be able
to select the wanted option. This is happening because the table row in
this list is having a css value for position property which is relative.

This was working on 15.0 but refactoring the views to the new owl
web client was the reason to break this because we change things
in the list renderer like removing the table-responsive div
from the view.

Desired behavior after PR is merged:
The view has been adjusted by removing the css style for the data-row
in document list view.

opw-3682703

closes odoo/enterprise#56650

X-original-commit: 5093ef4b04f6f0e16a6ec1bc50fe57b32f1493f2
Signed-off-by: Youssef Bashandy (yoba) <yoba@odoo.com>