[FIX] sale_external_tax: allow recomputation of taxes on confirmed SOs
Before this, clicking the "Compute Taxes" button on a confirmed sale
order does nothing. It was possible in previous versions to
re-calculate taxes on confirmed sale orders and it's especially handy
on subscriptions. Furthermore, it's possible to manually change the
taxes on confirmed sale orders anyway, so there's no reason to prevent
changing them through an external tax integration.

opw-3694079

closes odoo/enterprise#55004

X-original-commit: 05e8e081f6dc838cc9be8b77dc9efc490650a08b
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: Joren Van Onder (jov) <jov@odoo.com>