[FIX] account_reports: xlsx format
XLSX are now formatted automatically, both
in width and height.
We were applying text_wrap by default on them,
but it may be too much for automatic formatting.

To ensure that we won't impact negatively the result,
it was decided to revert that and limit ourselves to
handle the col width and to not tough the row height.

Task id # 3679470

closes odoo/enterprise#55065

X-original-commit: 650620536aaf1f4be63fe98c0a1e2da0146e4f9c
Signed-off-by: John Laterre (jol) <jol@odoo.com>