[IMP] whatsapp: add variable button
Add a variable button on the whatsapp template body to facilitate
the addition of new variables without the need for the user to enter
them manually.

Clicking on the button will automatically add the next variable at the
cursor position or at the end of the body if there's no cursor.
A popover will be displayed in case the maximum of 10 variables are
present in the template.

Task-3628187

closes odoo/enterprise#52396

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>