[FIX] knowledge: mark clipboard html as odooHtml
There was an issue with the `copy` feature of the `/clipboard` command. When
pasted, the content was formatted as if it was not originally `Odoo` content
(and some style features were removed, like the font color).

To remedy that, instead of using the `clipboard` API, the deprecated
`execCommand('copy')` function is used so that the `copy` handler of the
`OdooEditor` is properly triggered and the html content is properly set under
the `text/odoo-editor` custom "MIME type".

Remark: it is not possible use that "fake" "MIME type" to write data directly to
the clipboard because it is not officially recognized (will produce an error).

task-3700875

closes odoo/enterprise#56448

X-original-commit: 8ec8398430459c0410e20331088e71225ab2ed00
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>