[IMP] stock_barcode: scan multiple barcodes
This commit adds a new system parameter:
`stock_barcode.barcode_separator_regex`

Like its name is saying, it's a regex, and its default value will catch
commats, semicolons and linebreaks.

In the barcode App, when a barcode is scanned and it can be splitted
regarding the regex, the scanned barcode will be split into multiple
barcodes and each part will be processed independently.
Because a lot of barcode can be processed like that, the UI is blocked
while it's ongoing.

The purpose of this feature is to scan a bunch of serial/lot numbers at
once.

task-3595212

closes odoo/enterprise#55036

Related: odoo/odoo#150877
Related: odoo/upgrade#5740
Signed-off-by: Steve Van Essche <svs@odoo.com>