[IMP] account_reports: add deductibility columns to journal report
In some countries the journal report is used as a section of the tax
report to provide details about the computed tax amounts.

One such detail is the split between the deductible and non-deductible
part of the tax, as well as the tax due.

In order to provide these details, we add three new columns to the tax
overview in the journal report that indicate the amounts of the tax
that are deductible, non-deductible or due.

Also, we add a Global Tax Summary that spans all journals.

task-3696656

closes odoo/enterprise#57596

X-original-commit: 2aae2c10677180b979712a333aadcb22b3fcb50a
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>
Signed-off-by: Dylan Kiss (dyki) <dyki@odoo.com>