[FIX] account_avatax: stop calculating negative credit notes
Before this commit, credit notes would be calculated to have a
negative total amount which is incorrect and prevents posting.

New in Odoo 17 is that account_external_tax sets line.price_subtotal
on invoices. This was necessary to correctly support price-included
taxes for Brazil. A side-effect is that it uncovers an inconsistency
in the US integration [1]. Avatax returns negative amounts (both
lineAmount and tax) when the record is a ReturnInvoice. In Odoo we
don't invert the price_* fields on account.move.line for credit notes
so this commit inverts the sign coming from Avatax in these cases.

[1] PS. in https://github.com/odoo/enterprise/pull/45095#issuecomment-1696102292

opw-3699206

closes odoo/enterprise#55200

X-original-commit: 33fb031b8f0a500c9b47d56503a62106c64c135f
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: Joren Van Onder (jov) <jov@odoo.com>