[FIX] *: impossible to remove applications' user groups
Without this change it's impossible to remove the user groups from internal
users, because they're being re-added upon the module updates.

The reason is because there's a mechanism in place that synchronizes the newly
added groups to the default user, so they're automatically added to all internal
users.

For this reason and by convention, groups added to the default user should be
non-updateable, so system administrators can remove them if they wanted to.

closes odoo/enterprise#75177

X-original-commit: ec942d2ecc52663b94c2479e25f4b6fafcc47103
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>