[FIX] documents: stop guessing document partner
We revert the PRs below because it generates issues in the accounting
application when the ocr is detecting someone else as contact, and it cannot
overrule the one we detected in Documents:

In odoo/enterprise#39292, we have implemented:
If not explicitly defined on the documents.share, when receiving a document by
email, we set the contact to the partner corresponding to the email sender. If
no partner exists with that email, a new one is created.

In odoo/enterprise#50379, we have partially reverted the change:
When receiving a document from an unknown contact, we no longer create a new
partner. However, if the contact is known, we assign the already existing
partner to the document.

We revert here the last part that is to assign the author of the message to the
document if the author is a known partner. However, we keep the tests introduced
in odoo/enterprise#39292 with some minor adaptations.

How to reproduce:
- install documents with demo data
- send an email with an attachment to the inbox alias (that create a document
when receiving an email) from an email of an existing partner. It can be
simulated using odoo_mail_gate.py.
- find the document received and select it
- in the document inspector, the contact should be empty but it is not

Task-3672520

closes odoo/enterprise#56138

X-original-commit: 4398f3347fea00551f19fe28e6ab3c1fab389d42
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>
Signed-off-by: Pierre-Yves Dufays (pydu) <pydu@odoo.com>