[FIX] website_helpdesk: fix website menu url to avoid error
If website_menu don't have a url and website_helpdesk module is
installed, hence, we checked particular website_menu's url as a result we got
the false value for url if it's null, that's why  while connecting the database
"500 internal server error" is raised.

To resolve this issue, This PR will help to fetch only those records which have
url in website_menu.

closes odoo/enterprise#55037

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>