[FIX] web_studio: effect attr on a button should be an object.
On a CRM lead in studio, activate the rainbow man on the "Won" button.
Change the image.

Before this commit, there was a crash, because activating the rainbow man put the effect
attribute on the button to "true".
Adding an image on top of that crashed because the code expected an object.

Given the empirical specs at action_service.js:doActionButton and in effect_service.js, the "effect"
attribute doesn't have the expected effect if it is "true". We choose then to set the "effect" attribute on a node
- to "False" when it is supposed to be deactivated,
- a empty object "{}" if it is to be truthy, but displaying defaults. This object is enriched when customizing
details in studio.

After this commit, there is no crash anymore and one can customize the image in the rainbow man.

opw-3692982

closes odoo/enterprise#55086

X-original-commit: 7365a2589073a16be10726275b7c65f03c2c9b86
Signed-off-by: Luca Vitali (luvi) <luvi@odoo.com>