[FIX] account_sepa: use NURG SvcLvl code on generic ISO20022 payment
We used to skip the SvcLvl node on generic ISO20022 payments. This caused some banks to reject the XML file generated for them (though this node is not supposed to be mandatory). We now always set in on them, with the default NURG code, indicating that the payment is non-urgent.

OPW 3679650

closes odoo/enterprise#56320

X-original-commit: b03a06ee6be99b84c0bed02f705be03506f79449
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: William André (wan) <wan@odoo.com>