[IMP] room: use room id in bus notification types
Purpose:
--------
The bus service uses a single websocket connection that is shared across
multiple tabs. Therefore, sending a notification to one tab actually sends
it to every tab.
Currently, the notification type is the same regardless of which room
sends it. Therefore if 2 (or more) different room frontend views are
opened at the same time, they will both handle the same notification,
even though it was only concerning one of the rooms, leading to
inconsistent states.

This commit adds the room id in the bus notifications types.
This allows for the roomBookingView component to only listen for the
notifications intended for its room.

Task-3698135

closes odoo/enterprise#55039

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>