[FIX] planning,project_forecast: only compute template_id when it is needed
Before this commit, when the `allocated_hours` is altered in a `planning.slot`,
it will trigger the compute of `template_id` in `planning_slot` to finally do
nothing since the `allocated_hours` is not used inside that compute.
The problem by letting that field in the dependencies is the `compute_role_id`
will be triggered because the template_id potentially changed thanks to its
compute method and so the compute method of `slot_properties` will be triggered
and could raise an access error in the case the `planning.slot` cannot be read.

This commit removes `allocated_hours` field in the dependencies of
`compute_template_id` since it does not seem to have any reason to trigger
that compute when the allocated_hours is changed on a `planning.slot`.

closes odoo/enterprise#55261

X-original-commit: cd0d29f8491118d7ae136eb7f222de6463b2c15c
Related: odoo/odoo#151392
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>