[REF] pos*: unify owl app instantiation
pos*: pos_order_tracking_display, pos_preparation_display

There are multiple owl apps in the pos ecosystem.
For each app we follow a similar pattern of instantiation.
This means that code is repeated, which has often lead to
a worse developer experience.

In this commit we create a common function that performs
all the steps needed when starting an owl app and we adapt
the codebase to use it.

closes odoo/enterprise#55062

Task: 3701400
Related: odoo/odoo#150921
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>