[FIX] account_accountant,account_reports: branch user fixes
When a user only has access to a branch of a company, some basic flow
are broken by some AccessError or unaccessible parent's field (when it
should be) and/or other general errors because of the domain setup.

This commit aims to fix those behaviors.
This commit accompanies the AccessError and unaccessible parent fixes
from the odoo PR version.

closes odoo/enterprise#55345

Task-id: 3551305
Related: odoo/odoo#151562
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>
Signed-off-by: Yosua Nicolaus (yoni) <yoni@odoo.com>