[IMP] hr_payroll: headover report
This commit adds a new type of report to the payroll module. This ne reporting
allows to get the empployees that have a contract that was running on a given
date. The purpose of this report is to be able have a clear view of the
employees headcount of the company on a given date.

Task-3691392

Part-of: odoo/enterprise#55533
Co-authored-by: tivisse <yti@odoo.com>