[FIX] web_studio: ViewEditor: primary button invisible are readable
Have some buttons with class ".btn .btn-primary" and that are invisible in the arch.
In studio, click on "Show invisible elements".

The text of those buttons were not readable on the screen because of the lack of contrast.

This commit fixes this, the background becoming stripped on the basis of the the enterprise primary color.

closes odoo/enterprise#55657

X-original-commit: 5a9d2dfef5c81587eaa2ebe187093025a0811315
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>