[FIX] web_studio: ViewEditor: primary button invisible are readable
Have some buttons with class ".btn .btn-primary" and that are invisible in the arch.
In studio, click on "Show invisible elements".

The text of those buttons were not readable on the screen because of the lack of contrast.

This commit fixes this, the background becoming stripped while keeping the original background-
color set on the element, since the background-image property is used with a stripped repeating
mask.

closes odoo/enterprise#55730

X-original-commit: dfdece14dd57c35b95686ab4192eef7bd9d1b704
Signed-off-by: Luca Vitali (luvi) <luvi@odoo.com>
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>