[FIX] l10n_ke_hr_payroll: fix reports
Commit 7ad17cb introduced a change in the way
payroll reports are generated.
However, this change forgot to include those
changes in the kenyan l10n reporting.

This commit adapts the P9 form so it can be
generated again.

This commit also fixes NSSF and NHIF as the
default month was using a wrong calculation,
generating a traceback on January.

Finally, this commit introduces additional demo
data for the l10n with some payslips and
a P9 report.

closes odoo/enterprise#55206

Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>