[IMP] planning_*: remove working_days_count
Before this commit, the field 'working_days_count' was utilized in
reports related to scheduling. However, this field did not accurately
represent the number of working days for a resource. For instance, if a
resource has multiple non-overlapping shifts in a week, the calculation
is accurate. But, as soon as shifts overlap, the value no longer
reflects reality. Why? Because the number of working days is calculated
for each time slot based on the daily hours in the calendar,
independently of other shifts. This means, for a 40-hour week calendar
with 8-hour days, if there are two shifts from Monday to Friday, each
with 40 allocated hours, the working_days_count equals 10, which is
incorrect. It also doesn't work for employees with flexible hours due to
the absence of a standard calendar.

This commit removes the field as there is no viable solution to resolve
these issues.

taskid:3186627

closes odoo/enterprise#56060

Related: odoo/upgrade#5654
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>