[IMP] l10n_lt_reports: use only tags instead of tags / code mix
Currently the BS and P&L use a mix of account codes and tags.
But there is no fixed or legally required CoA in Lithuania.
So we should rather only use tags so that the users can more decide
where the accounts go in the BS / P&L.

This commit converts the codes to tags.
The P&L lines already using the domain engine for tags
were converted to account_codes engine.
Also the profit and loss entries in the BS are now computed the
same way as the P&L.

task-3701151

closes odoo/enterprise#56825

Related: odoo/odoo#154407
Related: odoo/upgrade#5706
Signed-off-by: William André (wan) <wan@odoo.com>