[FIX] account_reports: Change xpath for sale report code filters
Before this commit, the filter for EC Sale List codes
was invisible if the report doesn't have a variant.
It was caused by the xpath based on a div which can
be not present in the DOM. When this div is not present,
the filter doesn't appear.

no task id

closes odoo/enterprise#55187

X-original-commit: ebac41762888facab8e13938347e266d5bb2b659
Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>