[FIX] appointment_hr: fix traceback when accessing gantt view
Make sure to use the resource calendar of the employee.
Indeed, it is possible that the resource calendar set on the user
is different than the one on the employee. Which can leads to a
traceback if the resource calendar was not in the
unavailabilities_by_calendar.

We now use the employee resource calendar everywhere instead of the calendar
on the user.

task-3698586

closes odoo/enterprise#57844

X-original-commit: 40c12f87c97e06dd3acc53b5a1c9a54ae2c80d98
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Jérémy Hennecart (jeh) <jeh@odoo.com>