[FIX] appointment: include details for all appointments
The new design introduced in odoo/enterprise#41534 moved
the details inside a column used in the several templates
of the appointment flow, updating at each step.

However, it is not shown for anytime and custom
appointments. In order to avoid a large blank space
on the right of the screen and a strange design effect,
as well as to give the attendee the available details
as soon as they are available, always display the details
column.

However, as the duration may vary in the custom category,
do not show duration field at time selection for those
appointments. Show duration of selected slot on detail
screen.

Task-3701087

closes odoo/enterprise#55138

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>