[FIX] account_reports: handle duplicate aml id in general ledger
The key used was the id of the journal item, this is fine for most cases
but will fail due to duplicate if the "Cash basis" function is set and a
receivable aml is reconciled with two payment made at different dates. In this
case the report will show the receivable aml split in two whith the date/amount
of each payment.

To support this case, we now use a markup with the date of aml to produce
unique keys. When the two payments have the same date, it's ok since the
receivable aml is not split in the report.

Steps to reproduce
1/ Create an invoice
2/ Register 2 payments at different dates for this invoice
3/ General ledger with options "Unfold all" + "Cash basis"
=> Traceback
"Got duplicate key in t-foreach: ~account.account~6|~account.move.line~1128"

opw-3687800

closes odoo/enterprise#56338

X-original-commit: b62d9c2f6bdaffd955c8e6e67d0006f7d183e755
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Thomas Beckers (tbs) <tbs@odoo.com>