[IMP] knowledge: allow resequencing favorite children
In the favorite section of knowledge, it was possible to unfold a favorite and
see its children, but it was not possible to resequence them, and dragging them
would start a `drag` operation on their ancestor (which is set as a favorite).

Now it is possible to resequence them at the same level, and it will actually
resequence them in their respective section synchronously.

task-3555651

closes odoo/enterprise#55168

Related: odoo/odoo#151181
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>