[FIX] documents_spreadsheet: speed up list insertion tests
The helper `toggleCogMenuSpreadsheet`is quite slow atm.
As much as this should be investigated, there is a quick win by cutting
in the functional flow to insert a list in a spreadsheet and rely on the
implementation by triggering the bus event related to tje insert of the
list in a spreadsheet.

This revision speeds up the related concerned tests by ~6.2 seconds on a
local installation.

Testing the modules `documents_spreadsheet > list view` and
`documents_spreadsheet > insert_list_spreadsheet_menu`:

Average on 10 runs

closes odoo/enterprise#55461

Before: 11600 ms
After: 5400 ms
X-original-commit: d2586e5758a2698fd0989277d84b0dbaa4033018
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>