[IMP] web_grid, timesheet_grid: add line button
This commit modifies the criteria for showing the global "Add Line"
button, simplifying the process for users to add new entries.

In the "My Timesheets" section within the timesheet_grid module, the
"Add Line" button is now hidden when create_inline is true and display
no content is not displayed (that is, we keep the old generic behavior
in that menu). The reason why it is not always displayed in that menu is
because the incorporation of new lines is already streamlined through the
use of the timer feature.

taskid:3701294

closes odoo/enterprise#56037

Related: odoo/odoo#153395
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>