[IMP] timesheet_grid: hide future timesheets from the 'To Validate' menus
Before this commit, the timesheets that were dated in the future could
be visible in the "To Validate" menus. However, it is not possible
to validate such timesheets as they are dated in the future, which could
lead to confusion for the user.

To make things clear, this commit hides the timesheets that are dated in
the future from the "To Validate" menus.

task-3703516

closes odoo/enterprise#55233

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>