[FIX] stock_barcode: end tour properly
The tour try to delete a stock move line then validate the picking.
The issue is that the validate button is trigger before being sure it
available in the view.

This commit ensure the stock move line form view is closed after the
call to "Delete" before validating the picking.

closes odoo/enterprise#55239

Signed-off-by: William Henrotin (whe) <whe@odoo.com>