[FIX] project_enterprise: group expand portal user
This commit fixes the group_expand in porject_entreprise when searching
for a assignees (thus on the res.users model). The problem was that
the portal users were also displayed, which makes no sense as they
cannot be assigned to tasks.

related: https://github.com/odoo/odoo/pull/172973
task-3703637

closes odoo/enterprise#56038

Related: odoo/odoo#172973
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>